+++ vue/src/core/services/Deploy/models/DeploymentCronJob.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 29-08-2025.
+ * Time: 10:53.
+ */
+import {DeploymentCronJobDefinition} from './definitions/DeploymentCronJobDefinition';
+
+export class DeploymentCronJob extends DeploymentCronJobDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}